Javadoc implementation

+++ back/src/main/java/com/openclassrooms/mddapi/controller/UserController.java
@@ -1,12 +1,12 @@
 package com.openclassrooms.mddapi.controller;
 
+import java.util.Objects;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,36 +17,38 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.openclassrooms.mddapi.mapper.UserMapper;
 import com.openclassrooms.mddapi.model.User;
-import com.openclassrooms.mddapi.payload.JwtResponse;
 import com.openclassrooms.mddapi.payload.MessageResponse;
 import com.openclassrooms.mddapi.payload.UserUpdateRequest;
-import com.openclassrooms.mddapi.security.JwtUtil;
-import com.openclassrooms.mddapi.security.UserDetailsImpl;
-import com.openclassrooms.mddapi.security.UserDetailsServiceImpl;
 import com.openclassrooms.mddapi.service.UserService;
 
+/**
+ * User Controller
+ * @author Guillaume Belaud
+ * @version 0.1
+ */
 @RestController
 @RequestMapping("api/user")
 public class UserController {
 	
+	/**
+	 * @see com.openclassrooms.mddapi.mapper.UserMapper.java
+	 */
 	@Autowired
 	private UserMapper mapper;
 	
+	/**
+	 * @see com.openclassrooms.mddapi.service.UserService.java
+	 */
 	@Autowired
-	private UserService service;
+	private UserService service;	
 	
-	@Autowired
-	private AuthenticationManager authManager;
-	
-	@Autowired
-	private JwtUtil jwtUtil;
-	
-	@Autowired
-	private PasswordEncoder encoder;
-	
-	@Autowired
-	private UserDetailsServiceImpl userDetailsService;
-	
+	/**
+	 * Get a user details by id
+	 * @param id		string corresponding to a user
+	 * @return			httpResponse status 200 with user Dto object
+	 * 					else return status 400 for a NumberFormatException
+	 * 					or status 404 if id is unknown
+	 */
 	@GetMapping("/{id}")
     public ResponseEntity<?> findById(@PathVariable("id") String id) {
         try {
@@ -62,6 +64,14 @@ public class UserController {
         }
     }
 	
+	/**
+	 * Update a user details by id
+	 * @param id		string corresponding to a user
+	 * @return			httpResponse status 200 with successful message response
+	 * 					else return status 400 for a NumberFormatException
+	 * 					or status 404 if id is unknown
+	 * 					or status 401 if authenticate user is different to updating user
+	 */
 	@PutMapping("/{id}")
     public ResponseEntity<?> updateById(@PathVariable("id") String id, @RequestBody UserUpdateRequest updatedUser) {
         try {
@@ -69,6 +79,12 @@ public class UserController {
 
             if (user == null) {
                 return ResponseEntity.notFound().build();
+            }
+            
+            UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+            if(!Objects.equals(userDetails.getUsername(), user.getEmail())) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
             }
             
             user.setUserName(updatedUser.getUserName());
@@ -80,7 +96,16 @@ public class UserController {
             return ResponseEntity.badRequest().build();
         }
     }
-
+	
+	
+	/**
+	 * Delete a user by id
+	 * @param id		string corresponding to a user
+	 * @return			httpResponse status 200
+	 * 					else return status 400 for a NumberFormatException
+	 * 					or status 404 if id is unknown
+	 * 					or status 401 if authenticate user is different to deleting user
+	 */
     @DeleteMapping("{id}")
     public ResponseEntity<?> delete(@PathVariable("id") String id) {
         try {
@@ -90,7 +115,11 @@ public class UserController {
                 return ResponseEntity.notFound().build();
             }
 
-            //TODO : Add authentication verification before deleting
+            UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+            if(!Objects.equals(userDetails.getUsername(), user.getEmail())) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            }
 
             service.delete(Long.parseLong(id));
             return ResponseEntity.ok().build();
